Setup native (based on Tokenizer) `Reflector`
Previously, `roave/better-reflection` was used to retrieve all class names
(FQCN) from a file. As the roave library is wrapping this in an own
ReflectionClass object, it results in a lot of overhead. With big codebases
this can be a big difference in processing time.

Therefore, using the native PhpToken (Tokenizer; since PHP8.1), is faster.

(benchmark: 12K files, roave: ~12s, native: ~2s)